% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFunList.R
\name{getFunList}
\alias{getFunList}
\title{getFunList Function}
\usage{
getFunList(fn = "")
}
\arguments{
\item{#love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to
}
\examples{
getFunList()
}
\keyword{Model}
