% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{optim}
\alias{optim}
\title{optim Function}
\usage{
optim(
  par,
  fn,
  gr = NULL,
  ...,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  lower = -Inf,
  upper = Inf,
  control = list(),
  hessian = FALSE,
  smart = TRUE,
  gn.args = list()
)
}
\arguments{
\item{#love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to
}
\examples{
optim()
}
\keyword{Model}
