% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{testing}
\alias{testing}
\title{testing Function}
\usage{
testing(fn = "", verbose = list())
}
\arguments{
\item{#love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to
}
\examples{
testing()
}
\keyword{Model}
